requirements.txt
# Canvas MPC - Consolidated Requirements
# All dependencies for local development and production

# ============================================
# Core MCP & Canvas
# ============================================
mcp>=1.0.0
canvasapi>=2.0.0
python-dotenv>=1.0.0

# ============================================
# Backend - FastAPI & Server
# ============================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pydantic>=2.0.0

# ============================================
# Google APIs (Calendar & Gmail)
# ============================================
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# ============================================
# Database & Authentication (Optional)
# ============================================
# Supabase is optional - only needed if using auth features
supabase>=2.0.0

# ============================================
# Frontend - Streamlit
# ============================================
streamlit>=1.28.0
streamlit-option-menu>=0.3.6
requests>=2.31.0

# ============================================
# Utilities & Data Processing
# ============================================
# All standard library modules used:
# - asyncio (stdlib)
# - datetime, timedelta, timezone (stdlib)
# - zoneinfo (stdlib Python 3.9+)
# - json, os, sys, re (stdlib)
# - typing, functools, contextlib (stdlib)
# - email, base64 (stdlib)
# - html (stdlib)

# ============================================
# Development & Testing (Optional)
# ============================================
# Uncomment for development:
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# flake8>=6.0.0
# black>=23.0.0
# mypy>=1.0.0

# ============================================
# Production Extras (Optional)
# ============================================
# Uncomment for production monitoring:
# prometheus-client>=0.16.0
# sentry-sdk>=1.40.0
